# backend/openapi_check.json
{"openapi":"3.1.0","info":{"title":"Unclutr.ai API","description":"\nBackend for Unclutr.ai - The Decision & Control Layer for D2C.\n\n### 🔐 Authorization\nTo test protected endpoints, click the **Authorize** button:\n\n1. **Development Login**: Use any username and password `admin`.\n2. **Email/Password**: Use your real Firebase credentials. (Requires `FIREBASE_API_KEY` in `.env`)\n3. **Google Auth**: \n   - Login on the [Frontend](http://localhost:3000).\n   - Open DevTools (F12) -> Application -> Local Storage.\n   - Look for `firebase:authUser:...`.\n   - Copy the `stsTokenManager.accessToken`.\n   - Paste it directly into the **Bearer** field in Swagger (click Authorize -> JWT (Bearer)).\n","version":"v0.1.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login For Access Token","description":"OAuth2 compatible token login. \n1. Checks for Swagger Dev Password (admin).\n2. If not dev, attempts real Firebase Auth (Email/Password).","operationId":"login_for_access_token_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/sync":{"post":{"tags":["Auth"],"summary":"Sync User Endpoint","description":"Syncs Firebase User to Local DB. \nCall this after Firebase Login on Frontend.","operationId":"sync_user_endpoint_api_v1_auth_sync_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/health/full":{"get":{"tags":["Health"],"summary":"Health Check Full","description":"Comprehensive Health Check for the Developer Dashboard.\nChecks: Database, Firebase Admin, API Status.","operationId":"health_check_full_api_v1_health_full_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/onboarding/state":{"get":{"tags":["Onboarding"],"summary":"Get Onboarding Status","description":"Get current onboarding state and resume information.\n\nFrontend uses this to:\n1. Check if onboarding is complete\n2. Determine which page to resume from\n3. Hydrate form fields with saved data","operationId":"get_onboarding_status_api_v1_onboarding_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/onboarding/save":{"post":{"tags":["Onboarding"],"summary":"Save Progress","description":"Save onboarding progress for a specific page.\nCalled by 'Save & Exit' button or when navigating between pages.\n\nDoes NOT mark onboarding as complete - use /finish for that.","operationId":"save_progress_api_v1_onboarding_save_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveProgressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/onboarding/finish":{"post":{"tags":["Onboarding"],"summary":"Finish Onboarding","description":"Complete onboarding and commit all data to production models.\n\nCalled when user clicks 'Go to Dashboard' on finish page.\nCreates: Company, Brand, Workspace, Integrations.\nMarks onboarding as complete.\n\nIdempotent: Safe to call multiple times.","operationId":"finish_onboarding_api_v1_onboarding_finish_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/onboarding/step":{"post":{"tags":["Onboarding"],"summary":"Save Step","description":"DEPRECATED: Use /save endpoint instead.\nKept for backward compatibility.\n\nMaps step numbers to page names:\n- Step 1 → basics\n- Step 2 → channels\n- Step 3 → stack\n- Step 4 → finish","operationId":"save_step_api_v1_onboarding_step_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StepData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/dev/events":{"get":{"tags":["Dev"],"summary":"Stream Events","operationId":"stream_events_api_v1_dev_events_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/dev/reset-onboarding":{"post":{"tags":["Dev"],"summary":"Reset Onboarding","description":"Resets onboarding state for the current user.","operationId":"reset_onboarding_api_v1_dev_reset_onboarding_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/dev/purge-data":{"post":{"tags":["Dev"],"summary":"Purge Data","description":"Hard purge of all data associated with the current user's companies.\nDANGER: Dev only.","operationId":"purge_data_api_v1_dev_purge_data_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/datasources/":{"get":{"tags":["Datasources"],"summary":"Read Datasources","description":"Retrieve available datasources.","operationId":"read_datasources_api_v1_datasources__get","security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/DataSourceCategory"},{"type":"null"}],"title":"Category"}},{"name":"is_common","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Common"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataSource"},"title":"Response Read Datasources Api V1 Datasources  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasources/request":{"post":{"tags":["Datasources"],"summary":"Request Datasource","description":"Submit a request for a missing datasource.","operationId":"request_datasource_api_v1_datasources_request_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataSourceRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/datasources/requests":{"get":{"tags":["Datasources"],"summary":"Read Requests","description":"List all datasource requests (Internal/Control Tower use).","operationId":"read_requests_api_v1_datasources_requests_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DataSourceRequest"},"type":"array","title":"Response Read Requests Api V1 Datasources Requests Get"}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/overview":{"get":{"tags":["Metrics"],"summary":"Get Metrics Overview","description":"Get overview metrics for the control tower dashboard.\nReturns high-level KPIs and recent activity.","operationId":"get_metrics_overview_api_v1_metrics_overview_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/business":{"get":{"tags":["Metrics"],"summary":"Get Business Metrics","description":"Get business metrics for the specified number of days.\nReturns daily KPIs including users, companies, workspaces, and integrations.","operationId":"get_business_metrics_api_v1_metrics_business_get","security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/onboarding":{"get":{"tags":["Metrics"],"summary":"Get Onboarding Metrics","description":"Get onboarding funnel metrics.\nReturns completion rates, drop-off points, and average time per step.","operationId":"get_onboarding_metrics_api_v1_metrics_onboarding_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/integrations":{"get":{"tags":["Metrics"],"summary":"Get Integration Metrics","description":"Get integration health metrics.\nReturns sync success rates, error rates, and performance metrics.","operationId":"get_integration_metrics_api_v1_metrics_integrations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/customers":{"get":{"tags":["Metrics"],"summary":"Get Customer Metrics","description":"Get list of customers with their metrics.\nReturns company information, member counts, and health scores.","operationId":"get_customer_metrics_api_v1_metrics_customers_get","security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/users":{"get":{"tags":["Metrics"],"summary":"Get User Metrics","description":"Get user engagement metrics for the specified number of days.\nReturns daily engagement scores and activity metrics.","operationId":"get_user_metrics_api_v1_metrics_users_get","security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":90,"minimum":1,"default":7,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/export/customers":{"get":{"tags":["Metrics"],"summary":"Export Customers Csv","description":"Export customer data as CSV.","operationId":"export_customers_csv_api_v1_metrics_export_customers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}]}},"/api/v1/metrics/calculate/daily":{"post":{"tags":["Metrics"],"summary":"Calculate Daily Metrics","description":"Manually trigger calculation of daily metrics.\nUseful for backfilling or recalculating metrics.","operationId":"calculate_daily_metrics_api_v1_metrics_calculate_daily_post","security":[{"HTTPBearer":[]},{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"target_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BasicsData":{"properties":{"companyName":{"type":"string","maxLength":100,"minLength":1,"title":"Companyname","description":"Name of the legal entity"},"brandName":{"type":"string","maxLength":100,"minLength":1,"title":"Brandname","description":"Name of the brand"},"category":{"type":"string","maxLength":50,"minLength":1,"title":"Category","description":"D2C category"},"region":{"$ref":"#/components/schemas/RegionData"}},"type":"object","required":["companyName","brandName","category","region"],"title":"BasicsData"},"Body_login_for_access_token_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_v1_auth_login_post"},"ChannelsData":{"properties":{"selectedChannels":{"items":{"type":"string"},"type":"array","title":"Selectedchannels","description":"List of selected sales channel IDs"}},"type":"object","title":"ChannelsData"},"DataSource":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"},"category":{"$ref":"#/components/schemas/DataSourceCategory"},"website_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website Url"},"logo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Logo Url"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"auth_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Auth Method"},"is_active":{"type":"boolean","title":"Is Active","default":true},"is_coming_soon":{"type":"boolean","title":"Is Coming Soon","default":false},"is_common":{"type":"boolean","title":"Is Common","default":false},"is_implemented":{"type":"boolean","title":"Is Implemented","default":false},"theme_color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Theme Color","default":"#000000"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","slug","category"],"title":"DataSource"},"DataSourceCategory":{"type":"string","enum":["Storefront","Marketplace","QuickCommerce","Network","SocialCommerce","Marketing","Logistics","Payment","Accounting","Analytics","Communication","Retention"],"title":"DataSourceCategory"},"DataSourceRequest":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","title":"User Id"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"user_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Name"},"name":{"type":"string","title":"Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"status":{"$ref":"#/components/schemas/RequestStatus","default":"PENDING"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["user_id","name"],"title":"DataSourceRequest"},"FinishData":{"properties":{},"type":"object","title":"FinishData"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RegionData":{"properties":{"country":{"type":"string","title":"Country"},"currency":{"type":"string","title":"Currency"},"timezone":{"type":"string","title":"Timezone"},"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"flag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Flag"}},"type":"object","required":["country","currency","timezone"],"title":"RegionData"},"RequestCreate":{"properties":{"name":{"type":"string","title":"Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},"type":"object","required":["name"],"title":"RequestCreate"},"RequestStatus":{"type":"string","enum":["PENDING","REVIEWED","APPROVED","REJECTED"],"title":"RequestStatus"},"SaveProgressRequest":{"properties":{"page":{"type":"string","enum":["basics","channels","stack","finish"],"title":"Page"},"data":{"anyOf":[{"$ref":"#/components/schemas/BasicsData"},{"$ref":"#/components/schemas/ChannelsData"},{"$ref":"#/components/schemas/StackData"},{"$ref":"#/components/schemas/FinishData"},{"additionalProperties":true,"type":"object"}],"title":"Data"}},"type":"object","required":["page","data"],"title":"SaveProgressRequest"},"StackData":{"properties":{"selectedTools":{"items":{"type":"string"},"type":"array","title":"Selectedtools","description":"List of selected tool IDs"}},"type":"object","title":"StackData"},"StepData":{"properties":{"step":{"type":"integer","title":"Step"},"data":{"additionalProperties":true,"type":"object","title":"Data"}},"type":"object","required":["step","data"],"title":"StepData","description":"DEPRECATED: Use SaveProgressRequest instead"},"UserRead":{"properties":{"email":{"type":"string","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"picture_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Picture Url"},"is_active":{"type":"boolean","title":"Is Active","default":true},"is_superuser":{"type":"boolean","title":"Is Superuser","default":false},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_login_at":{"type":"string","format":"date-time","title":"Last Login At"},"onboarding_completed":{"type":"boolean","title":"Onboarding Completed","default":false}},"type":"object","required":["email","id","created_at","last_login_at"],"title":"UserRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"},"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}